requirement.txt
Django==3.0.1
djangorestframework==3.11.0
django-cors-middleware==1.5.0
django-oauth-toolkit==1.2.0
mysqlclient==1.4.6
django-filter==2.2.0
